requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
openai>=1.0.0
matplotlib>=3.8.0
tabulate>=0.9.0
python-multipart>=0.0.9
httpx>=0.27.0
